put widgets in modules in ratatui-widgets crate
Rather than exporting all the widgets at the root for ratatui-widgets,
we now put them in modules. This makes the crate more organized and
easier to navigate. For instance the table Cell and Row types are now
ratatui_widgets::table::Cell and ratatui_widgets::table::Row.

The widgets maintain their existing locations in the ratatui crate
(flat re-exports at ratatui::widgets).

One small reorganization is that block::BorderType is now
borders::BorderType.
