+++ ratatui-widgets/src/list.rs
@@ -1,3 +1,5 @@
+//! The [`List`] widget is used to display a list of items and allows selecting one or multiple
+//! items.
 pub use self::{
     item::ListItem,
     list::{List, ListDirection},
